% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraire_statuts_versions.R
\name{extraire_statuts_versions}
\alias{extraire_statuts_versions}
\title{Fonction qui extrait les statuts pour chacun des fichiers taxref contenu dans un répertoire}
\usage{
extraire_statuts_versions(repertoire)
}
\arguments{
\item{repertoire}{Texte. Chemin vers le répertoire où se trouvent les fichiers TAXREF.}
}
\value{
Un dataframe avec les calculs des nombres et pourcentages d'EEE par groupe taxonomique
    pour chacune des versions de TAXREF.
}
\description{
Fonction qui extrait les statuts pour chacun des fichiers taxref contenu dans un répertoire
}
\examples{
\dontrun{
resultats <- extraire_statuts_versions(repertoire = "raw_data")
}

}
